requirements.txt
python-telegram-bot[webhooks]==20.5
google-api-python-client==2.92.0
google-auth==2.23.2
Flask==2.2.5
apscheduler==3.10.1
python-dateutil==2.8.2
requests==2.31.0
SQLAlchemy==2.0.20
alembic==1.12.0
psycopg2-binary==2.9.7
python-json-logger==2.0.4
gunicorn==20.1.0
pytest==7.0.1
pytest-asyncio==0.16.0
pytest-cov==3.0.0